third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/icon-allowed.html

<!DOCTYPE html>
<html>
<meta http-equiv="Content-Security-Policy" content="img-src http://localhost"/>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    testRunner.dumpIconChanges();
}

function createLink(rel, src) {
    var link = document.createElement('link');
    link.rel = rel;
    link.href = src;
    document.head.appendChild(link);
}

window.addEventListener('DOMContentLoaded', function() {
    createLink('icon', 'http://localhost/foo?q=from_icon');
    if (window.testRunner) {
        testRunner.notifyDone();
    }
});
</script>
<p>Use callbacks to show that favicons are loaded as allowed by CSP when link tags are dynamically added to the page.</p>
</html>
