third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/icon-allowed-expected.txt

main frame - didChangeIcons
Use callbacks to show that favicons are loaded as allowed by CSP when link tags are dynamically added to the page.
